modules/locations-event-api/requirements.txt
google-cloud==0.34.0
grpcio==1.59.3
grpcio-tools==1.59.3
kafka-python==2.0.2
protobuf==4.25.1